pyproject.toml
[tool.poetry]
name = "kbdgen"
version = "2.0.0-alpha.0"
description = "A generator for keyboard layouts"
authors = ["Brendan Molloy <brendan@bbqsrc.net>"]
license = "Apache-2.0"
readme = "README.rst"
homepage = "https://github.com/divvun/kbdgen"
documentation = "https://divvun.github.io/kbdgen"
keywords = ["keyboard", "generator", "cldr"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "License :: OSI Approved :: Apache Software License"
]
include = [
    "kbdgen/*.yaml",
    "kbdgen/gen/bin/*",
    "LICENSE",
    "CHANGELOG.md",
]

[tool.poetry.scripts]
kbdgen = "kbdgen.__main__:main"
cldr2kbdgen = "kbdgen.cldr:cldr2kbdgen_main"
kbdgen2cldr = "kbdgen.cldr:kbdgen2cldr_main"

[tool.poetry.dependencies]
python = "^3.4"
requests = "^2.20.1"
humanize = "^0.5.1"
pyyaml = "^3.13"
lcid = "^1.0.1"
lxml = "^4.2.5"
langcodes = "^1.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
